Refacto + start enriching with more stuff
