+++ src/SerilogWeb.Classic.WebApi/Classic/WebApi/PreApplicationStartModule.cs
@@ -9,7 +9,7 @@ namespace SerilogWeb.Classic.WebApi
         {
             GlobalConfiguration.Configuration.Services.Add(typeof(IExceptionLogger),
                 new WebApiExceptionLogger());
-            GlobalConfiguration.Configuration.Filters.Add(new StoreWebApInfoInHttpContext());
+            GlobalConfiguration.Configuration.Filters.Add(new StoreWebApInfoInHttpContextActionFilter());
         }
     }
 }
