+++ src/SerilogWeb.Classic.WebApi/Classic/WebApi/WebApiRequestInfoKey.cs
@@ -8,6 +8,8 @@ namespace SerilogWeb.Classic.WebApi
 {
     enum WebApiRequestInfoKey
     {
-        UrlTemplate
+        UrlTemplate,
+        ActionName,
+        ControllerName
     }
 }
