+++ src/SerilogWeb.Classic.WebApi/SerilogWeb.Classic.WebApi.csproj
@@ -83,7 +83,7 @@
     <Compile Include="Classic\WebApi\Enrichers\HttpContextCurrent.cs" />
     <Compile Include="Classic\WebApi\Enrichers\WebApiUrlTemplateEnricher.cs" />
     <Compile Include="Classic\WebApi\PreApplicationStartModule.cs" />
-    <Compile Include="Classic\WebApi\StoreWebApInfoInHttpContext.cs" />
+    <Compile Include="Classic\WebApi\StoreWebApInfoInHttpContextActionFilter.cs" />
     <Compile Include="Classic\WebApi\WebApiExceptionLogger.cs" />
     <Compile Include="Classic\WebApi\WebApiRequestInfoKey.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
